+++ lib/feature/quiz/infrastructure/entity/hitting_stats.dart
