+++ lib/feature/quiz/infrastructure/entity/hitting_stats.freezed.dart
