+++ lib/feature/quiz/infrastructure/entity/supabase_hitter.dart
