+++ lib/feature/quiz/infrastructure/entity/supabase_hitter.freezed.dart
