+++ lib/main.dart
@@ -29,8 +29,8 @@ import 'feature/loading/application/loading_state.dart';
 import 'feature/loading/presentation/loading_widget.dart';
 import 'feature/quiz/application/hitter_quiz_state.dart';
 import 'feature/quiz/domain/hitter_repository.dart';
-import 'feature/quiz/infrastructure/supabase/supabase_hitter_repository.dart';
-import 'feature/quiz/infrastructure/supabase/supabase_providers.dart';
+import 'feature/quiz/infrastructure/supabase_hitter_repository.dart';
+import 'feature/quiz/infrastructure/supabase_providers.dart';
 import 'feature/quiz_result/application/quiz_result_state.dart';
 import 'feature/quiz_result/domain/quiz_result_repository.dart';
 import 'feature/quiz_result/infrastructure/firebase_quiz_result_repository.dart';
